+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added new `Object::findOne()` method to search for and return a single child tree from an object. 
 - Added new argument `returnFirstFound` to `Query::search`, used to implement the above. 
-- 
+- Fixed some compiler warnings. 
 
 ## 1.3.1 * 2024-03-29